Rename files and namespaces. Delete old files
