--- PhotoOrganizerLib.Tests/TestFiles/testdata-template.json
@@ -1,8 +1,0 @@
-{
-    "FileName": "FILE.jpg",
-    "Expected": {
-        "HashValue": "",
-        "DateTime": "",
-        "ISO": 0
-    }
-}
